requirements.txt
# Core ML frameworks
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Visualization & utilities
matplotlib>=3.7.0
tqdm>=4.65.0

# ML & evaluation
scikit-learn>=1.3.0
transformers>=4.30.0

# Diarization & metrics
pyannote.core>=5.0.0
pyannote.metrics>=3.0.0
pyannote.audio>=3.0.0

# Speech recognition & timestamping
openai-whisper>=20230314
whisper-timestamped>=1.13.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
